sprawdzanie odbicia piłki i odbicie od paletki

+++ ArkanoEgo/Classes/Booster.cs
@@ -20,8 +20,8 @@ namespace ArkanoEgo.Classes
     {
         public int rad { get; set; } // promień kulki (potrzebny do odbić)
         public int speed { get; set; }
-        public int posX { get; set; } // - w góre, + w dół
-        public int posY { get; set; } // - w lewo, + w prawo
+        public double posX { get; set; } // - w góre, + w dół
+        public double posY { get; set; } // - w lewo, + w prawo
 
         public bool top { get; set; } // top = true; bottom = false;
         public bool left { get; set; } // left = true; right = false;
